requirements.txt
beautifulsoup4
blas
brotli-python
bzip2
ca-certificates
certifi
cffi
chardet
charset-normalizer
click
colorama
contourpy
cycler
deep-translator
et-xmlfile
filelock
fonttools
freetype
fsspec
googletrans
h11
h2
httpcore
hpack
hstspreload
httpx
huggingface-hub
hyperframe
idna
intel-openmp
jinja2

joblib
kiwisolver
lcms2
lerc
libblas
libcblas
libcublas
libcufft
libcurand
libcusolver
libcusparse
libdeflate
libexpat
libffi
libhwloc
libiconv
libjpeg-turbo
liblapack
libnpp
libnvjpeg
libpng
libsqlite
libtiff
libwebp-base
libxml2
libzlib
m2w64-gcc-libgfortran
m2w64-gcc-libs
m2w64-gcc-libs-core
m2w64-gmp
m2w64-libwinpthread-git
markupsafe
matplotlib
mkl
mpmath
msys2-conda-epoch
networkx
nltk
numpy
openjpeg
openpyxl
openssl
packaging
pandas
pillow
pip
protobuf
pthread-stubs
pthreads-win32
pycparser
pyparsing
pysocks
pytz
pyyaml
regex
requests
rfc3986
safetensors
sentencepiece
setuptools
six
sniffio
soupsieve
sympy
tbb
tk
tokenizers
tqdm
transformers
triton
typing_extensions
tzdata
ucrt
urllib3
vc
vc14_runtime
vs2015_runtime
wheel
win_inet_pton
xformers==0.0.27.post2
xorg-libxau
xorg-libxdmcp
xz
yaml
zstandard
zstd
